Major rework to common components and navigation
