+++ package.json
@@ -18,6 +18,7 @@
     "eslint-config-next": "13.4.19",
     "next": "13.4.19",
     "react": "18.2.0",
+    "react-cookie": "^6.1.1",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
