+++ pages/librarian.tsx
@@ -5,6 +5,8 @@ import {Button, Form, Input, message, Modal, Select, Space, Table, Tag} from "an
 import {User} from ".prisma/client";
 import {Book} from ".prisma/client";
 import {Checkout} from ".prisma/client";
+import AddBook from "./shared/addBook"
+import Logout from './shared/logout';
 const inter = Inter({ subsets: ['latin'] })
 
 const layout = {
@@ -16,31 +18,9 @@ const tailLayout = {
     wrapperCol: { offset: 8, span: 12 },
 };
 
-export default function Home() {
-    const [users, setUsers] = useState<User[]>([]);
+export default function Librarian() {
     const [books, setBooks] = useState<Book[]>([]);
     const [checkouts, setCheckouts] = useState<Checkout[]>([]);
-    const [isModalOpen, setIsModalOpen] = useState(false);
-    const [form] = Form.useForm();
-
-    const onFinish = async (values: any) => {
-        console.log(values);
-        setIsModalOpen(false);
-        fetch('/api/create_book', {
-            method: 'POST',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(values)
-        }).then(async response => {
-            if (response.status === 200) {
-                const book = await response.json();
-                message.success('Created book: ' + book.title);
-            } else message.error(
-                `Failed to create book:\n ${JSON.stringify(await response.json())}`);
-        }).catch(res=>{message.error(res)})
-    };
 
     const onDelete = async (book: any) => {
         console.log(book);
@@ -138,21 +118,6 @@ export default function Home() {
         },
     ];
 
-
-    const onReset = () => {
-        form.resetFields();
-    };
-
-    const showModal = () => {
-        setIsModalOpen(true);
-        form.resetFields();
-    };
-
-    const handleCancel = () => {
-        setIsModalOpen(false);
-        form.resetFields();
-    };
-
     useEffect(()=>{
         fetch('api/all_books', {method: "GET"})
             .then(res => {
@@ -175,47 +140,8 @@ export default function Home() {
     if (!checkouts) return "Give me a second";
 
     return  <>
-        <Button type="primary" onClick={showModal}>
-            Add Book
-        </Button>
-        <Modal title="Add Book" onCancel={handleCancel}
-               open={isModalOpen} footer={null}  width={800}>
-            <Form
-                {...layout}
-                form={form}
-                name="control-hooks"
-                onFinish={onFinish}
-                style={{ maxWidth: 600 }}
-            >
-                <Form.Item name="isbn" label="ISBN" rules={[{ required: true }]}>
-                    <Input />
-                </Form.Item>
-                <Form.Item name="title" label="Title" rules={[{ required: true }]}>
-                    <Input />
-                </Form.Item>
-                <Form.Item name="author" label="Author" rules={[{ required: true }]}>
-                    <Input />
-                </Form.Item>
-                <Form.Item name="copies" label="Copies" rules={[{ required: true }]}>
-                    <Input type="number"/>
-                </Form.Item>
-                <Form.Item name="copiesOut" label="Copies Checked Out" rules={[{ required: true }]}>
-                    <Input type="number"/>
-                </Form.Item>
-
-                <Form.Item {...tailLayout} >
-                    <Button type="primary" htmlType="submit">
-                        Submit
-                    </Button>
-                    <Button htmlType="button" onClick={onReset}>
-                        Reset
-                    </Button>
-                    <Button  htmlType="button" onClick={handleCancel}>
-                        Cancel
-                    </Button>
-                </Form.Item>
-            </Form>
-        </Modal>
+        <AddBook />
+        <Logout />
 
         <Table columns={colBooks} dataSource={books} />;
         <Table columns={colCheckouts} dataSource={checkouts} />;
